Develop

+++ backendCompteurApp/src/main/java/com/compteurapp/backendcompteurapp/DTO/FactureDTO.java
@@ -1,9 +1,9 @@
 package com.compteurapp.backendcompteurapp.DTO;
 
-import com.compteurapp.backendcompteurapp.model.FactureStatement;
+import com.compteurapp.backendcompteurapp.enums.FactureStatement;
 
 public class FactureDTO {
-    public int prix;
+    public double prix;
     public FactureStatement etat;
     public Long idCompteurData;
 }
